1.非全部处理时，限制处理数量值从文件中读取 2.关注信息，去除yes|no, using ok. 3.读取字符串的值变更为unsigned int

+++ codesets/base/string/cstringpub.cpp
@@ -350,9 +350,9 @@ bool CStringPub::strSimLenZero(QString str)
     return (str.simplified().length() == 0);
 }
 
-int CStringPub::strSimLen(QString str)
+unsigned int CStringPub::strSimLen(QString str)
 {
-    return str.simplified().length();
+    return (unsigned int)str.simplified().length();
 }
 
 QString CStringPub::strSim(QString str)
